+++ src/App.js
@@ -121,8 +121,14 @@ import SettingLayout from "./NEFDashboard/containers/settingLayout";
 import SettingPage from "./NEFDashboard/containers/settings";
 import Products from "./NEFDashboard/containers/products";
 import CreateInvoice from "./NEFDashboard/containers/createInvoiceDashboard";
+import AuthLayout from "./NEFDashboard/containers/authLayout";
+import LoginForm from "./NEFDashboard/containers/login";
+import SignForm from "./NEFDashboard/containers/signUp";
+import ForgotPassword from "./NEFDashboard/containers/forgotPassword";
+import SetPasswordForm from "./NEFDashboard/containers/setPassword";
 import NewLanding, { NewB2B, NewB2C, NewResources } from "./pages/Landing";
 import Resources from "./components/resources";
+import PageNotFound from "./NEFDashboard/components/pageNotFound";
 import { Helmet } from "react-helmet";
 import { useTranslation } from "react-i18next";
 
@@ -220,6 +226,24 @@ function App() {
                 >
                   <ScrollToTop>
                     <Routes>
+                      <Route path="/login" element={<LoginForm />} />
+                      <Route path="/sign-up" element={<SignForm />} />
+                      <Route
+                        path="/forgot-password"
+                        element={
+                          <AuthLayout>
+                            <ForgotPassword />
+                          </AuthLayout>
+                        }
+                      />
+                      <Route
+                        path="/set-password"
+                        element={
+                          <AuthLayout>
+                            <SetPasswordForm />
+                          </AuthLayout>
+                        }
+                      />
                       <Route
                         path="/new-settings"
                         element={
@@ -569,6 +593,15 @@ function App() {
                             {/* <Navigation /> */}
                             <Pay />
                             {/* <Footer /> */}
+                          </>
+                        }
+                      />
+                      <Route
+                        path="*"
+                        element={
+                          <>
+                            <Navigation />
+                            <PageNotFound />
                           </>
                         }
                       />